Fix: arena page screens
