+++ src/components/AddItem/AddItem.js
@@ -37,11 +37,11 @@ const AddItem = () => {
   };
 
   return (
-    <div className="row add-item">
-      <header style={{ marginTop: '7%' }}>
+    <div className="add-item">
+      <header style={{ paddingTop: '100px' }}>
         <h1 className="app-title">Add An Arena</h1>
       </header>
-      <div className="d-flex p-0 flex-column justify-content-center align-items-center  item-arena">
+      <div className="d-flex p-0 flex-column justify-content-center align-items-center">
         <form onSubmit={(e) => onSubmit(e)} className="mt-5 mx-3 ">
           <input type="text" className="form-control m-1" id="name" placeholder="Enter Name" name="name" onChange={(e) => handleChange(e)} style={divStyle} />
           <textarea rows="4" cols="50" type="text" className="form-control m-1" id="description" placeholder="Enter Description" name="description" onChange={(e) => handleChange(e)} style={divStyle} />
