+++ src/components/Logo.jsx
@@ -1,0 +1,3 @@
+const Logo = () => <span className="brackets" />;
+
+export default Logo;
