+++ src/utils/itemsHelper.js
@@ -1,0 +1,11 @@
+const calculateItemsPerView = (screenWidth) => {
+  if (screenWidth > 768) {
+    if (screenWidth <= 991) {
+      return 2;
+    }
+    return 3;
+  }
+  return 1;
+};
+
+export default calculateItemsPerView;
